Update README.md
You are to code a SinglyLinkedList with head and tail references. A linked list is a collection of nodes, each having a data item and references to other nodes. In a SinglyLinkedList, each node has a reference to the next node. Since it is non-circular, the next reference for the tail is null.

Do not use a phantom node to represent the start or end of your list. A phantom or sentinel node is a node that does not store data held by the list and is used solely to indicate the start or end of a linked list. If your list contains n elements, then it should contain exactly n nodes.

The SinglyLinkedList must follow the requirements stated in the javadocs of each method you must implement. Your linked list implementation will use the default constructor (the one with no parameters) which is automatically provided by Java. Do not write your own constructor.

Nodes

The linked list consists of nodes. A class SinglyLinkedList is provided to you. This class has getter and setter methods to access and mutate the structure of the nodes.

Adding

You will implement three add() methods. One will add to the front, one will add to the back, and one will add to anywhere in the list given a specific index. See the javadocs for more details.

Removing

You will also implement three remove() methods – from the front, the back, or anywhere in the list given a specific index. Make sure that there is no longer any way to access the removed node so that the node will be garbage collected. See the javadocs for more details.